Move evaluate functions to evaluate module
