+++ src/senselab/audio/tasks/bioacoustic_qc/__init__.py
@@ -5,13 +5,17 @@ from .bioacoustic_qc import (
     activity_to_evaluations,
     activity_to_taxonomy_tree_path,
     check_quality,
+    subtree_to_evaluations,
+)
+from .evaluate import (
     evaluate_audio,
     evaluate_batch,
     evaluate_dataset,
-    subtree_to_evaluations,
+    get_metric,
 )
 
 __all__ = [
+    "get_metric",
     "subtree_to_evaluations",
     "activity_to_dataset_taxonomy_subtree",
     "check_quality",
